+++ src/commonMain/kotlin/io/github/unredundant/neonctl/commands/BranchCommands.kt
@@ -6,9 +6,14 @@ import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.required
 import io.github.unredundant.neonctl.models.BranchCreateRequest
 import io.github.unredundant.neonctl.models.BranchCreateRequestEndpointOptions
+import io.github.unredundant.neonctl.models.BranchesResponse
 import io.github.unredundant.neonctl.models.CreatedBranch
+import io.github.unredundant.neonctl.models.Endpoint
 import io.github.unredundant.neonctl.models.EndpointType
+import io.github.unredundant.neonctl.models.EndpointsResponse
 import io.github.unredundant.neonctl.requests.createProjectBranch
+import io.github.unredundant.neonctl.requests.listProjectBranches
+import io.github.unredundant.neonctl.requests.listProjectEndpoints
 import io.github.unredundant.neonctl.util.ErrorMessage
 import io.github.unredundant.neonctl.util.NeonCtlUtils
 import io.ktor.client.HttpClient
@@ -95,8 +100,46 @@ class BranchCommands(private val client: () -> HttpClient) : CliktCommand(
     name = "list"
   ) {
     private val config by findObject<BranchCommandConfig>()
-    override fun run() {
-      echo("List branches")
+    override fun run(): Unit = runBlocking {
+      val (projectId, client) = config ?: error("No config found")
+      val result = client.listProjectBranches(projectId)
+
+      when (result.status.value) {
+        in 200..299 -> {
+          val body: BranchesResponse = result.body()
+          echo(NeonCtlUtils.json.encodeToString(BranchesResponse.serializer(), body))
+        }
+
+        else -> {
+          val body: String = result.body()
+          echo(
+            NeonCtlUtils.json.encodeToString(
+              ErrorMessage.serializer(),
+              ErrorMessage(
+                message = "List branches failed",
+                code = result.status.value,
+                details = body
+              )
+            )
+          )
+        }
+      }
+    }
+  }
+
+  object GetPrimaryBranchEndpoint : CliktCommand(
+    help = "Returns the endpoint for the primary branch",
+    name = "primary_endpoint"
+  ) {
+    private val config by findObject<BranchCommandConfig>()
+    override fun run() = runBlocking {
+      val (projectId, client) = config ?: error("No config found")
+      val branches: BranchesResponse = client.listProjectBranches(projectId).body()
+      val primaryBranch = branches.branches.find { it.primary } ?: error("No primary branch found")
+      val endpoints: EndpointsResponse = client.listProjectEndpoints(projectId).body()
+      val primaryEndpoint = endpoints.endpoints.find { it.branchId == primaryBranch.id }
+        ?: error("No primary endpoint found")
+      echo(NeonCtlUtils.json.encodeToString(Endpoint.serializer(), primaryEndpoint))
     }
   }
 }