+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - May 21st, 2023
+
+### Added
+
+- List branches command
+- Get primary branch endpoint command
+
 ## [0.1.0] - May 20th, 2023
 
 ### Added